--- WDG_iOS/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  WDG_iOS
-//
-//  Created by 정찬웅 on 10/27/23.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
