+++ WDG_iOS/View/PostView.swift
@@ -1,0 +1,18 @@
+//
+//  PostView.swift
+//  WDG_iOS
+//
+//  Created by 정찬웅 on 10/28/23.
+//
+
+import SwiftUI
+
+struct PostView: View {
+    var body: some View {
+        Text("Post")
+    }
+}
+
+//#Preview {
+//    PostView()
+//}
