+++ WDG_iOS/View/TabView.swift
@@ -1,0 +1,18 @@
+//
+//  TabView.swift
+//  WDG_iOS
+//
+//  Created by 정찬웅 on 10/28/23.
+//
+
+import SwiftUI
+
+struct TabView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    TabView()
+}
