+++ WDG_iOS/WDG_iOSApp.swift
@@ -15,7 +15,7 @@ struct WDG_iOSApp: App {
     @UIApplicationDelegateAdaptor var appDelegate : AppDelegate
     var body: some Scene {
         WindowGroup {
-            MainListView()
+            ContentView()
         }
     }
 }
